Add support for loading SFRs from EC2
That is as opposed to loading via config files in S3 as we do right now.
To start with we will support loading via both until we can roll all the
SFRs to add the necessary tags.

The main motivation is that this will make it easier to deal with
cancelled running SFRs since we can load all the SFRs and just ask EC2
for their status rather than adding/removing S3 files via terraform.