Remove dns lookup for agent hostname to IP
Instead use a regex to convert the slave_pid to an IP. This is how the
paasta cluster autoscaler does this step. It is hacky but it seems to
be a lot more reliable than the DNS lookup which for some reason is
intermittently failing. The biggest risk is that mesos change the format
of this field in the future but at least that will be an easy error to
spot when we try to upgrade.