Change clusterman logging setup
I'm not really sure this is "right". But I'd like opinions/suggestions.

I don't think we should have get_clusterman_logger setting up handlers etc. So I
moved this to a setup func. My aim is that I see log messages from other
clusterman modules which is currently not the case. However this also
catches other non-clusterman modules such as boto which are noisy in
debug mode. Nothing here is set to debug mode but I think something in
the signals repos is...