Only process draining host once
If clusterman submits a host for draining and we are simply waiting for
a host to be terminated (whilst it is draining). Then there is nothing
stopping clusterman submitting the host again on another run if the time
between runs is less than the draining timeout. This ensures that the
drainer will only process hosts 1 for each submit->drain->terminate
cycle.

See PAASTA-15079 for context