Make processing list time based
This is better than simply keeping a set in memory and removing on
termination since in that time window clusterman may still have added
hosts to the draining queue which will then be processed again after a
successful termination.

Here we dedupe based on time and ignore duplicate instance_ids in a
certain time window.