Catch ValueError when loading multiple SFRs
The spot notification processing code must load all the
SpotFleetResourceGroups in the cluster so that it can check if
the spot notification is from a PaaSTA agent. This can sometimes fail if
clusterman does not have current instance data for a particular SFR i.e.
if it contains exotic new instance types. In this case it's okay to
continue and just load the other SFRs since worst case we will just miss
the spot notification and forgo draining.