Allow setting an AWS profile override
Right now, Clusterman will always use /etc/boto_cfg creds - which means
that we will not be able to access prod resources from devboxen
normally.

We can do one of two things:
* Use our personal permissions
* Route folks to boxes like adhoc-prod for running clusterman against
  prod clusters

We'd like to minimize our usage of adhoc-prod in general, so I've gone
for the first option and will be setting the relevant env vars in our
wrappers.